+++ guillotina/contrib/redis/dm.py
@@ -1,4 +1,5 @@
 from guillotina import configure
+from guillotina import metrics
 from guillotina.contrib.redis import get_driver
 from guillotina.files.adapter import DBDataManager
 from guillotina.interfaces import IExternalFileStorageManager
@@ -8,6 +9,31 @@ from guillotina.transactions import get_transaction
 
 import json
 import time
+
+
+try:
+    import prometheus_client
+
+    REDIS_OPS = prometheus_client.Counter(
+        "guillotina_dm_redis_ops_total",
+        "Total count of ops by type of operation and the error if there was.",
+        labelnames=["type", "error"],
+    )
+    REDIS_OPS_PROCESSING_TIME = prometheus_client.Histogram(
+        "guillotina_dm_redis_ops_processing_time_seconds",
+        "Histogram of operations processing time by type (in seconds)",
+        labelnames=["type"],
+    )
+
+    class watch(metrics.watch):
+        def __init__(self, operation: str):
+            super().__init__(
+                counter=REDIS_OPS, histogram=REDIS_OPS_PROCESSING_TIME, labels={"type": operation},
+            )
+
+
+except ImportError:
+    watch = metrics.watch  # type: ignore
 
 
 @configure.adapter(for_=IExternalFileStorageManager, provides=IUploadDataManager, name="redis")
@@ -23,7 +49,8 @@ class RedisFileDataManager(DBDataManager):
         if self._data is None:
             redis = await self.get_redis()
             key = self.get_key()
-            data = await redis.get(key)
+            with watch("get"):
+                data = await redis.get(key)
             if not data:
                 self._data = {}
             else:
@@ -42,7 +69,9 @@ class RedisFileDataManager(DBDataManager):
         redis = await self.get_redis()
         key = self.get_key()
         self._data["last_activity"] = time.time()
-        await redis.set(key, json.dumps(self._data, cls=GuillotinaJSONEncoder), expire=self._ttl)
+        value = json.dumps(self._data, cls=GuillotinaJSONEncoder)
+        with watch("set"):
+            await redis.set(key, value, expire=self._ttl)
 
     async def get_redis(self):
         if self._redis is None:
@@ -66,4 +95,5 @@ class RedisFileDataManager(DBDataManager):
         # and clear the cache key
         redis = await self.get_redis()
         key = self.get_key()
-        await redis.delete(key)
+        with watch("delete"):
+            await redis.delete(key)