+++ guillotina/contrib/redis/driver.py
@@ -6,6 +6,7 @@ except ImportError:
     raise
 
 from guillotina import app_settings
+from guillotina import metrics
 from guillotina.contrib.redis.exceptions import NoRedisConfigured
 from typing import Any
 from typing import List
@@ -14,6 +15,31 @@ from typing import Optional
 import asyncio
 import backoff
 import logging
+
+
+try:
+    import prometheus_client
+
+    REDIS_OPS = prometheus_client.Counter(
+        "guillotina_cache_redis_ops_total",
+        "Total count of ops by type of operation and the error if there was.",
+        labelnames=["type", "error"],
+    )
+    REDIS_OPS_PROCESSING_TIME = prometheus_client.Histogram(
+        "guillotina_cache_redis_ops_processing_time_seconds",
+        "Histogram of operations processing time by type (in seconds)",
+        labelnames=["type"],
+    )
+
+    class watch(metrics.watch):
+        def __init__(self, operation: str):
+            super().__init__(
+                counter=REDIS_OPS, histogram=REDIS_OPS_PROCESSING_TIME, labels={"type": operation},
+            )
+
+
+except ImportError:
+    watch = metrics.watch  # type: ignore
 
 
 logger = logging.getLogger("guillotina.contrib.redis")
@@ -45,10 +71,12 @@ class RedisDriver:
     @backoff.on_exception(backoff.expo, (OSError,), max_time=30, max_tries=4)
     async def _connect(self):
         settings = app_settings["redis"]
-        self._pool = await aioredis.create_pool(
-            (settings["host"], settings["port"]), **settings["pool"], loop=self._loop
-        )
-        self._conn = await self._pool.acquire()
+        with watch("create_pool"):
+            self._pool = await aioredis.create_pool(
+                (settings["host"], settings["port"]), **settings["pool"], loop=self._loop
+            )
+        with watch("acquire_conn"):
+            self._conn = await self._pool.acquire()
         self._pubsub_subscriptor = aioredis.Redis(self._conn)
 
     async def finalize(self):
@@ -72,25 +100,29 @@ class RedisDriver:
         args: List[Any] = []
         if expire is not None:
             args[:] = [b"EX", expire]
-        ok = await self._pool.execute(b"SET", key, data, *args)
+        with watch("set"):
+            ok = await self._pool.execute(b"SET", key, data, *args)
         assert ok == b"OK", ok
 
     async def get(self, key: str) -> str:
         if self._pool is None:
             raise NoRedisConfigured()
-        return await self._pool.execute(b"GET", key)
+        with watch("get"):
+            return await self._pool.execute(b"GET", key)
 
     async def delete(self, key: str):
         if self._pool is None:
             raise NoRedisConfigured()
-        await self._pool.execute(b"DEL", key)
+        with watch("delete"):
+            await self._pool.execute(b"DEL", key)
 
     async def delete_all(self, keys: List[str]):
         if self._pool is None:
             raise NoRedisConfigured()
         for key in keys:
             try:
-                await self._pool.execute(b"DEL", key)
+                with watch("delete_many"):
+                    await self._pool.execute(b"DEL", key)
                 logger.debug("Deleted cache keys {}".format(keys))
             except Exception:
                 logger.warning("Error deleting cache keys {}".format(keys), exc_info=True)
@@ -101,14 +133,16 @@ class RedisDriver:
         ops = [b"FLUSHDB"]
         if async_op:
             ops.append(b"ASYNC")
-        await self._pool.execute(*ops)
+        with watch("flush"):
+            await self._pool.execute(*ops)
 
     # PUBSUB API
 
     async def publish(self, channel_name: str, data: str):
         if self._pool is None:
             raise NoRedisConfigured()
-        await self._pool.execute(b"publish", channel_name, data)
+        with watch("publish"):
+            await self._pool.execute(b"publish", channel_name, data)
 
     async def unsubscribe(self, channel_name: str):
         if self._pubsub_subscriptor is None: