+++ guillotina/tests/fixtures.py
@@ -29,6 +29,7 @@ import aiohttp
 import asyncio
 import json
 import os
+import prometheus_client.registry
 import pytest
 
 
@@ -498,3 +499,12 @@ class DBUsersRequester(ContainerRequesterAsyncContextManager):
 @pytest.fixture(scope="function")
 async def dbusers_requester(guillotina):
     return DBUsersRequester(guillotina)
+
+
+@pytest.fixture(scope="function")
+async def metrics_registry():
+    for collector in prometheus_client.registry.REGISTRY._names_to_collectors.values():
+        if not hasattr(collector, "_metrics"):
+            continue
+        collector._metrics.clear()
+    yield prometheus_client.registry.REGISTRY